chore: move dir & rename component
