+++ feat/auth/src/main/java/com/emotionstorage/auth/ui/SignupCompleteScreen.kt
@@ -28,7 +28,7 @@ import com.emotionstorage.auth.presentation.SignupCompleteSideEffect
 import com.emotionstorage.auth.presentation.SignupCompleteViewModel
 import com.emotionstorage.domain.model.User.AuthProvider
 import com.emotionstorage.ui.component.button.CtaButton
-import com.emotionstorage.ui.component.TopAppBar
+import com.emotionstorage.ui.component.appBar.TopAppBar
 import com.emotionstorage.ui.theme.MooiTheme
 
 @Composable
