+++ feat/daily-report/src/main/java/com/emotionstorage/daily_report/ui/DailyReportDetailScreen.kt
@@ -35,7 +35,7 @@ import com.emotionstorage.daily_report.ui.component.DailyReportSummaries
 import com.emotionstorage.domain.model.DailyReport
 import com.emotionstorage.domain.model.DailyReport.EmotionLog
 import com.emotionstorage.ui.component.FullLoadingScreen
-import com.emotionstorage.ui.component.TopAppBar
+import com.emotionstorage.ui.component.appBar.TopAppBar
 import com.emotionstorage.ui.theme.MooiTheme
 import java.time.LocalDateTime
