+++ feat/my/src/main/java/com/emotionstorage/my/ui/NotificationSettingScreen.kt
@@ -33,7 +33,7 @@ import com.emotionstorage.my.ui.component.DayOfWeekSelector
 import com.emotionstorage.my.ui.component.ReminderTimeComponent
 import com.emotionstorage.my.ui.component.ToggleRow
 import com.emotionstorage.ui.component.bottomSheet.TimePickerBottomSheet
-import com.emotionstorage.ui.component.TopAppBar
+import com.emotionstorage.ui.component.appBar.TopAppBar
 import com.emotionstorage.ui.theme.MooiTheme
 import java.time.DayOfWeek
