+++ feat/my/src/main/java/com/emotionstorage/my/ui/TermsAndPrivacyScreen.kt
@@ -21,7 +21,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.emotionstorage.ui.content.PrivacyPolicyContent
 import com.emotionstorage.ui.content.TermsOfServiceContent
-import com.emotionstorage.ui.component.TopAppBar
+import com.emotionstorage.ui.component.appBar.TopAppBar
 import com.emotionstorage.ui.theme.MooiTheme
 
 @Composable
