+++ feat/my/src/main/java/com/emotionstorage/my/ui/WithDrawNoticeScreen.kt
@@ -22,7 +22,7 @@ import com.emotionstorage.my.presentation.MyPageSideEffect
 import com.emotionstorage.my.presentation.MyPageViewModel
 import com.emotionstorage.my.ui.component.WithDrawNoticeContent
 import com.emotionstorage.ui.component.Modal
-import com.emotionstorage.ui.component.TopAppBar
+import com.emotionstorage.ui.component.appBar.TopAppBar
 import com.emotionstorage.ui.component.button.CtaButton
 import com.emotionstorage.ui.component.button.CtaButtonType
 import com.emotionstorage.ui.theme.MooiTheme
