+++ feat/time-capsule-detail/src/main/java/com/emotionstorage/time_capsule_detail/ui/SaveTimeCapsuleScreen.kt
@@ -53,7 +53,7 @@ import com.emotionstorage.ui.component.toast.AppSnackbarHost
 import com.emotionstorage.ui.component.bottomSheet.DatePickerBottomSheet
 import com.emotionstorage.ui.component.FullLoadingScreen
 import com.emotionstorage.ui.component.toast.Toast
-import com.emotionstorage.ui.component.TopAppBar
+import com.emotionstorage.ui.component.appBar.TopAppBar
 import com.emotionstorage.ui.component.bottomSheet.YearMonthPickerBottomSheet
 import com.emotionstorage.ui.theme.MooiTheme
 import com.emotionstorage.ui.util.subBackground
