+++ feat/tutorial/src/main/java/com/emotionstorage/tutorial/ui/onBoarding/PrivacyPolicyDetailScreen.kt
@@ -9,7 +9,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import com.emotionstorage.ui.content.PrivacyPolicyContent
-import com.emotionstorage.ui.component.TopAppBar
+import com.emotionstorage.ui.component.appBar.TopAppBar
 import com.emotionstorage.ui.theme.MooiTheme
 
 @Composable
